Improve traits in model, in particular for reference handling

+++ openapi-generator-for-spring-api/src/main/java/de/qaware/openapigeneratorforspring/common/reference/fortype/ReferenceDeciderForType.java
@@ -1,7 +1,7 @@
 package de.qaware.openapigeneratorforspring.common.reference.fortype;
 
 public interface ReferenceDeciderForType<T> {
-    default boolean turnIntoReference(T item, int numberOfUsages) {
+    default boolean turnIntoReference(T item, long numberOfUsages) {
         return numberOfUsages > 1;
     }
 }
